log.c: Spray extra `FORMAT(...)' attributes to muffle Clang warnings.
Clang gets really upset about non-literal format strings, unless it can
check from the function attributes that you're playing by the rules.

I've taken the liberty of simplifying the annotations on static
functions: rather than duplicating the entire argument list, it suffices
to attach the attribute to the start of the function definition.

Signed-off-by: Mark Wooding <mdw@distorted.org.uk>
Acked-by: Ian Jackson <ijackson@chiark.greenend.org.uk>