style: util.[ch]: Introduce hex_encode_alloc name
Prompted by review of 7be31e47b2a8
  "util.[ch]: Factor out hex encoding and decoding utilities."
which says
  The interface is a bit odd, but it will fit with the uses
  I have in mind.

Not sure if it's the encode or decode interface which is referred to.
Certainly there should be a non-allocating variant.  I decided to
rename the allocating one.

The two separate buffer arguments to hex_decode are indeed a bit odd
but IMO tolerable.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>