site: Move main `struct msg' into site_incoming
We are going to want this in more places, and this is going to involve
threading it through site_incoming.  So make this a local variable
there, rather than in each of the process_msgN functions.

We rename the variable `named_msg' to `msg': it was called `named_msg'
because it was only valid after our calls to named_for_us, but now it
is valid after process_msgN too.

No overall functional change, except that stack usage is improved (by
removing a copy of struct msg).

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>