site: Pass msg into enter_new_state
The rules for when this is initialised, in site_incoming, are a
slightly complicated, so document them.

Examination of these rules reveals that the msg argument to
process_msg1 should be const, since process_msg1 (unlike the other
process_msgN functions) receives this, rather than generating it.

No functional change.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>