rsa1: rsapriv_apply: Introduce macros for cfgfatal* and free
We now make a distinction in rsapriv_apply's body between errors which
mean the file is not an rsa private key file, from other errors.

We replace free with a macro too because we are going to change the
error handling so as to support non-fatal early return.

No functional change.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>