rsa1: Provide rsa1_loadpriv function
We use the rsapriv_load_ctx infrastructure to provide a key loading
function suiting the scheme loadpriv interface.

I chose to use fmemopen because it is convenient.  We have an
autoconf-based portability fixup for eg BSDs (i.e., an implementation
in terms of funopen).

This is going to go in the schemes table.  It is not hooked in yet, so
won't be called.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>