site: Update peer keys from ~update file
We use the pathprefix_template facility for our public key set files.

Right now we document this behaviour in a comment in site.c, which is
the wrong place really.  We'll move it out later.

This commit introduces one spurious \n in a message, which will be
tidied up later.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>