make-secnet-sites: prefix names when writing sites file
SUMMARY

make-secnet-sites must copy names (vpn, location and site names) from
the input sites file (which is not wholly trusted) to the secnet
config file.

Since secnet 0.5.0 we check these name strings for syntactic sanity.

However, they still live in a common namespace.  When secnet is
evaluating a part of the config file, it looks up through the
containing structures.

It is possible for a sites file to specify, say, a site called "dh" or
something.  These currently generate bad configuration files.

IMPLEMENTATION

There are 4 lines of code which look like a nascent attempt to fix
this problem but (i) they are wrong and (ii) the reserved list is not
actually used anywhere.

Instead, we fix this by prefixing vpn names with V, location names
with L, and site names with S.  These obviously won't clash with
conventional names, predefined (builtin) closures, etc.  This is fine
because sensible configurations do not refer to the individual
make-secnet-sites config key names directly.

We achieve this by replacing the relevant references to the node's
name with calls to a new kname() method on the `level' class.  The new
behaviour is made optional, for compatibility reasons, but is enabled
by default.

We must disable this feature in the make-secnet-site tests because we
have a fixed expected output.  But the tests with actual secnet are
oblivious, since "stest: Use top-level mss-generated `all-sites' key".

THREAT ANALYSIS

An attacker can't do anything very interesting with this. right now.
A malicious sites file can generate a configuration file which secnet
will reject.  This is a nuisance but right now we don't have a system
for automatically incorporating sites data.  So I am not treating this
as a vulnerability.

It's not a more serious problem because the attacker can only bind the
keys to dictionaries containing site information written by
make-secnet-sites.  When the secnet configuration consumer code looks
something up and gets a dictionary when it was expecting something
else, it is an error.

There are few places where a dictionary is expected:

Path lookups expect a dictionary, but the only interesting scopes that
the attacker can pollute are the per-site config dictionaries, which
do not refer to paths, only individual keys.  (The "vpn" and
"all-sites" constructs from make-secnet sites contain path references,
but all the attacker can do there is to rebind what is supposed to be
an entry in the root namespace, resulting in a path which cannot be
resolved because it looks for too many subcomponents.)

There aren't currently any conventional keys with dictionary values
other than site's "comm-info" where the only key is
"dedicated-interface-addr".  The attacker could define a location
"comm-info" with a site "dedicated-interface-addr", which would be
interpreted as by peer sites as a "dedicated-interface-addr" key
referrinng to a dictionary (an error), or a site "comm-info" which
wouldn't contain that subkey at all and would therefore have no
effect.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>