make-secnet-sites: Delete old `reserved' attempt
This appears to be a nascent attempt to fix the problem now solved by
  "make-secnet-sites: prefix names when writing sites file"
by reserving the names of make-secnet-sites keywords.

However this is wrong because what needs to be reserved are secnet
config file conventional key names, and predefined words; there is no
clash between node names and make-secnet-sites keywords because they
appear in different parts of the sites file syntax (and the mss
keywords don't appear in the config file at all).

And anyway the `reserved' dictionary was never used anywhere so this
mechanism wasn't complete.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>