base91: Python: make `import' work
We have subtree merged the base91 python module as base91-python.

We will want to be able to `import base91'.  So make a symlink
`base91.py', and arrange to install it.  ($(INSTALL) installs a copy
of the file, not the link, as we need it to.)

This is getting rather intimate with the file structure of the
upstream base91-python.  If base91-python ever becomes more
complicated, and this breaks, it would surely be best to make it into
its own package.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>