base91: Build the C basE91 code with our own renaming
We want to use a variant character set.  This will mean patching the
library.  (That will happen in a moment.)  It seems unwise, and
perhaps rude, to make a thing that has the same name as the official
basE91 but has a different and incompatible encoding.

So we change all the names: `basE91' and `base91' become base91s.

We don't change the leaf filenames because that's rather
inconvenient.  We do put it all in a base91s subdirectory.

Rather than running the basE91 upstream makefile, or trying to sed it,
or something, we simply make our own makefile.  It is simple enough.
The only complication is that the "source" files are all actually made
by this name substitution perl rune.

Overall, the chief result is that we build a base91s/base91s utility,
by default.

Right now it does not have our charset change, so it is not ready for
use yet.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>