rsa: Pass an additional maybe_loc to load_err
Use it instead of l->loc if it is non-0.  But it isn't, yet, so no
functional change.

This will allow us to use load_err for handling errors in loadpub,
without regressing the error reporting: the individual bignums have
individual locs.  (Perhaps retaining this is gold-plating, but it is
not particularly hard.)

Rename the function to keep the table within 80 columns.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>