secnet: loadpub interface: Return a closure_t too
We are going to want this for the `make-public' verb.  In principle we
could just return the closure and the call site that wants the ops
could fish it out, but given that we have both this seems nicer.

(The secnet closure interface paradigm is not particularly pretty but
I don't fancy reworking it now...)

We will want this for loadpriv, too.  That will come (considerably)
later.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>