vcfgfatal_maybefile: New suffix argument
We are going to want to provide a log_if that calls *cfgfatal*.
Messages sent to a log_if do not have a \n, but messages passed to
cfgfatal do.  (This is daft, btw.)

To help resolve this impedance mismatch, provide a "suffix" parameter.
We only need this for vcfgfatal_maybefile, not for any of the other
variants.

No functional change.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>